use now stdint.h to be sure to use 2, 4 or 8 bytes when filling the windows structures

+++ pts/pe/peviewer/pestruct.h
@@ -1,6 +1,8 @@
 #ifndef PESTRUCT_H
 #define PESTRUCT_H
 
+#include <stdint.h>
+
 /* These defines describe the meanings of the bits in the Characteristics field */
 #define IMAGE_FILE_RELOCS_STRIPPED      0x0001 /* No relocation info */
 #define IMAGE_FILE_EXECUTABLE_IMAGE     0x0002
@@ -30,117 +32,117 @@
 #define IMAGE_NUMBEROF_DIRECTORY_ENTRIES 16
 
 typedef struct _IMAGE_DOS_HEADER {
-    unsigned short e_magic;      /* 00: MZ Header signature */
-    unsigned short e_cblp;       /* 02: Bytes on last page of file */
-    unsigned short e_cp;         /* 04: Pages in file */
-    unsigned short e_crlc;       /* 06: Relocations */
-    unsigned short e_cparhdr;    /* 08: Size of header in paragraphs */
-    unsigned short e_minalloc;   /* 0a: Minimum extra paragraphs needed */
-    unsigned short e_maxalloc;   /* 0c: Maximum extra paragraphs needed */
-    unsigned short e_ss;         /* 0e: Initial (relative) SS value */
-    unsigned short e_sp;         /* 10: Initial SP value */
-    unsigned short e_csum;       /* 12: Checksum */
-    unsigned short e_ip;         /* 14: Initial IP value */
-    unsigned short e_cs;         /* 16: Initial (relative) CS value */
-    unsigned short e_lfarlc;     /* 18: File address of relocation table */
-    unsigned short e_ovno;       /* 1a: Overlay number */
-    unsigned short e_res[4];     /* 1c: Reserved words */
-    unsigned short e_oemid;      /* 24: OEM identifier (for e_oeminfo) */
-    unsigned short e_oeminfo;    /* 26: OEM information; e_oemid specific */
-    unsigned short e_res2[10];   /* 28: Reserved words */
-    unsigned int e_lfanew;      /* 3c: Offset to extended header */
+    uint16_t e_magic;      /* 00: MZ Header signature */
+    uint16_t e_cblp;       /* 02: Bytes on last page of file */
+    uint16_t e_cp;         /* 04: Pages in file */
+    uint16_t e_crlc;       /* 06: Relocations */
+    uint16_t e_cparhdr;    /* 08: Size of header in paragraphs */
+    uint16_t e_minalloc;   /* 0a: Minimum extra paragraphs needed */
+    uint16_t e_maxalloc;   /* 0c: Maximum extra paragraphs needed */
+    uint16_t e_ss;         /* 0e: Initial (relative) SS value */
+    uint16_t e_sp;         /* 10: Initial SP value */
+    uint16_t e_csum;       /* 12: Checksum */
+    uint16_t e_ip;         /* 14: Initial IP value */
+    uint16_t e_cs;         /* 16: Initial (relative) CS value */
+    uint16_t e_lfarlc;     /* 18: File address of relocation table */
+    uint16_t e_ovno;       /* 1a: Overlay number */
+    uint16_t e_res[4];     /* 1c: Reserved words */
+    uint16_t e_oemid;      /* 24: OEM identifier (for e_oeminfo) */
+    uint16_t e_oeminfo;    /* 26: OEM information; e_oemid specific */
+    uint16_t e_res2[10];   /* 28: Reserved words */
+    uint32_t e_lfanew;      /* 3c: Offset to extended header */
 } IMAGE_DOS_HEADER, *PIMAGE_DOS_HEADER;
 
 typedef struct _IMAGE_DATA_DIRECTORY {
-    unsigned int VirtualAddress;
-    unsigned int Size;
+    uint32_t VirtualAddress;
+    uint32_t Size;
 } IMAGE_DATA_DIRECTORY, *PIMAGE_DATA_DIRECTORY;
 
 typedef struct _IMAGE_FILE_HEADER {
-    unsigned short Machine;
-    unsigned short NumberOfSections;
-    unsigned int TimeDateStamp;
-    unsigned int PointerToSymbolTable;
-    unsigned int NumberOfSymbols;
-    unsigned short SizeOfOptionalHeader;
-    unsigned short Characteristics;
+    uint16_t Machine;
+    uint16_t NumberOfSections;
+    uint32_t TimeDateStamp;
+    uint32_t PointerToSymbolTable;
+    uint32_t NumberOfSymbols;
+    uint16_t SizeOfOptionalHeader;
+    uint16_t Characteristics;
 } IMAGE_FILE_HEADER, *PIMAGE_FILE_HEADER;
 
 typedef struct _IMAGE_OPTIONAL_HEADER {
-    unsigned short Magic;
+    uint16_t Magic;
     unsigned char MajorLinkerVersion;
     unsigned char MinorLinkerVersion;
-    unsigned int SizeOfCode;
-    unsigned int SizeOfInitializedData;
-    unsigned int SizeOfUnitializedData;
-    unsigned int AddressOfEntryPoint;
-    unsigned int BaseOfCode;
-    unsigned int BaseOfData;
-    unsigned int ImageBase;
-    unsigned int SectionAlignment;
-    unsigned int FileAlignment;
-    unsigned short MajorOperatingSystemVersion;
-    unsigned short MinorOperatingSystemVersion;
-    unsigned short MajorImageVersion;
-    unsigned short MinprImageVersion;
-    unsigned short MajorSubsystemVersion;
-    unsigned short MinorSubsystemVersion;
-    unsigned int Win32VersionValue;
-    unsigned int SizeOfVersion;
-    unsigned int SizeOfHeaders;
-    unsigned int CheckSum;
-    unsigned short Subsystem;
-    unsigned short DllCharacteristics;
-    unsigned short SizeOfStackReserve;
-    unsigned short SizeOfStackCommit;
-    unsigned short SizeOfHeapReserve;
-    unsigned short SizeOfHeapCommit;
-    unsigned short LoaderFlags;
-    unsigned short NumberOfRvaAndSizes;
+    uint32_t SizeOfCode;
+    uint32_t SizeOfInitializedData;
+    uint32_t SizeOfUnitializedData;
+    uint32_t AddressOfEntryPoint;
+    uint32_t BaseOfCode;
+    uint32_t BaseOfData;
+    uint32_t ImageBase;
+    uint32_t SectionAlignment;
+    uint32_t FileAlignment;
+    uint16_t MajorOperatingSystemVersion;
+    uint16_t MinorOperatingSystemVersion;
+    uint16_t MajorImageVersion;
+    uint16_t MinprImageVersion;
+    uint16_t MajorSubsystemVersion;
+    uint16_t MinorSubsystemVersion;
+    uint32_t Win32VersionValue;
+    uint32_t SizeOfVersion;
+    uint32_t SizeOfHeaders;
+    uint32_t CheckSum;
+    uint16_t Subsystem;
+    uint16_t DllCharacteristics;
+    uint16_t SizeOfStackReserve;
+    uint16_t SizeOfStackCommit;
+    uint16_t SizeOfHeapReserve;
+    uint16_t SizeOfHeapCommit;
+    uint16_t LoaderFlags;
+    uint16_t NumberOfRvaAndSizes;
     IMAGE_DATA_DIRECTORY DataDirectory[IMAGE_NUMBEROF_DIRECTORY_ENTRIES];
 } IMAGE_OPTIONAL_HEADER, *PIMAGE_OPTIONAL_HEADER;
 
 typedef struct _IMAGE_OPTIONAL_HEADER64 {
-    unsigned short Magic;
+    uint16_t Magic;
     unsigned char MajorLinkerVersion;
     unsigned char MinorLinkerVersion;
-    unsigned int SizeOfCode;
-    unsigned int SizeOfInitializedData;
-    unsigned int SizeOfUninitializedData;
-    unsigned int AddressOfEntryPoint;
-    unsigned int BaseOfCode;
-    unsigned int long ImageBase;
-    unsigned int SectionAlignment;
-    unsigned int FileAlignment;
-    unsigned short MajorOperationgSystemVersion;
-    unsigned short MinorOperatingSystemVersion;
-    unsigned short MajorImageVersion;
-    unsigned short MinorImageVersion;
-    unsigned short MajorSubsystemVersion;
-    unsigned short MinorSubsystemVersion;
-    unsigned int Win32VersionValue;
-    unsigned int SizeOfImage;
-    unsigned int SizeOfHeaders;
-    unsigned int CheckSum;
-    unsigned short Subsystem;
-    unsigned short DllCharacteristics;
-    unsigned int long SizeOfStackReserve;
-    unsigned int long SizeOfStackCommit;
-    unsigned int long SizeOfHeapReserve;
-    unsigned int long SizeOfHeapCommit;
-    unsigned int LoaderFlags;
-    unsigned int NumberOfRvaAndSizes;
+    uint32_t SizeOfCode;
+    uint32_t SizeOfInitializedData;
+    uint32_t SizeOfUninitializedData;
+    uint32_t AddressOfEntryPoint;
+    uint32_t BaseOfCode;
+    uint64_t ImageBase;
+    uint64_t SectionAlignment;
+    uint32_t FileAlignment;
+    uint16_t MajorOperationgSystemVersion;
+    uint16_t MinorOperatingSystemVersion;
+    uint16_t MajorImageVersion;
+    uint16_t MinorImageVersion;
+    uint16_t MajorSubsystemVersion;
+    uint16_t MinorSubsystemVersion;
+    uint32_t Win32VersionValue;
+    uint32_t SizeOfImage;
+    uint32_t SizeOfHeaders;
+    uint32_t CheckSum;
+    uint16_t Subsystem;
+    uint16_t DllCharacteristics;
+    uint64_t SizeOfStackReserve;
+    uint64_t SizeOfStackCommit;
+    uint64_t SizeOfHeapReserve;
+    uint64_t SizeOfHeapCommit;
+    uint32_t LoaderFlags;
+    uint32_t NumberOfRvaAndSizes;
     IMAGE_DATA_DIRECTORY DataDirectory[IMAGE_NUMBEROF_DIRECTORY_ENTRIES];
 } IMAGE_OPTIONAL_HEADER64, *PIMAGE_OPTIONAL_HEADER64;
 
 typedef struct _IMAGE_NT_HEADERS {
-    unsigned int Signature;
+    uint32_t Signature;
     IMAGE_FILE_HEADER FileHeader;
     IMAGE_OPTIONAL_HEADER OptionalHeader;
 } IMAGE_NT_HEADERS, *PIMAGE_NT_HEADERS;
 
 typedef struct _IMAGE_NT_HEADERS64 {
-    unsigned int Signature;
+    uint32_t Signature;
     IMAGE_FILE_HEADER FileHeader;
     IMAGE_OPTIONAL_HEADER64 OptionalHeader;
 } IMAGE_NT_HEADERS64, *PIMAGE_NT_HEADERS64;
@@ -150,17 +152,17 @@ typedef struct _IMAGE_NT_HEADERS64 {
 typedef struct _IMAGE_SECTION_HEADER {
     unsigned char Name[IMAGE_SIZEOF_SHORT_NAME];
     union {
-        unsigned int PhysicalAddress;
-        unsigned int VirtualSize;
+        uint32_t PhysicalAddress;
+        uint32_t VirtualSize;
     } Misc;
-    unsigned int VirtualAddress;
-    unsigned int SizeOfRawData;
-    unsigned int PointerToRawData;
-    unsigned int PointerToRelocations;
-    unsigned int PointerToLinenumbers;
-    unsigned short NumberOfRelocations;
-    unsigned short NumberOfLinenumbers;
-    unsigned int Characteristics;
+    uint32_t VirtualAddress;
+    uint32_t SizeOfRawData;
+    uint32_t PointerToRawData;
+    uint32_t PointerToRelocations;
+    uint32_t PointerToLinenumbers;
+    uint16_t NumberOfRelocations;
+    uint16_t NumberOfLinenumbers;
+    uint32_t Characteristics;
 } IMAGE_SECTION_HEADER, *PIMAGE_SECTION_HEADER;
 
 #endif /* PESTRUCT_H */
